feat: show changed file contents in UncommittedChangesDialog
