+++ src/Designer/frontend/packages/shared/src/components/GiteaHeader/VersionControlButtons/components/FileChangesTable/FileChangesTable.module.css
@@ -1,0 +1,8 @@
+.table {
+  width: 100%;
+}
+
+.table .fileStatusCell {
+  vertical-align: top;
+  width: 1%;
+}
