+++ src/Designer/frontend/packages/shared/src/components/GiteaHeader/VersionControlButtons/components/FileChangesTable/FilePath/FilePath.test.tsx
@@ -4,7 +4,7 @@ import type { FilePathProps } from './FilePath';
 import { FilePath } from './FilePath';
 import type { ServicesContextProps } from 'app-shared/contexts/ServicesContext';
 import { queriesMock } from 'app-shared/mocks/queriesMock';
-import { renderWithProviders } from '../../../../../../mocks/renderWithProviders';
+import { renderWithProviders } from '../../../../mocks/renderWithProviders';
 
 const fileNameMock = 'fileName.json';
 const filePathWithoutNameMock = 'mock/file/path/to';
