+++ src/Designer/frontend/packages/shared/src/components/GiteaHeader/VersionControlButtons/components/FileChangesTable/index.ts
@@ -1,0 +1,2 @@
+export { FileChangesTable } from './FileChangesTable';
+export type { FileChangesTableProps } from './FileChangesTable';
