+++ src/Designer/frontend/packages/shared/src/components/GiteaHeader/VersionControlButtons/components/ShareChangesPopover/CommitAndPushContent/FileChangesInfoModal/FileChangesInfoModal.module.css
@@ -7,15 +7,6 @@
   width: var(--studio-file-changes-modal-width);
 }
 
-.dialog .table {
-  width: 100%;
-}
-
-.dialog .table .fileStatusCell {
-  vertical-align: top;
-  width: 1%;
-}
-
 .openDialogButton {
   margin-bottom: var(--fds-spacing-4);
 }
